requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
flask-marshmallow==0.11.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
html5lib==1.0.1
importlib-metadata==1.5.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
marshmallow==3.4.0
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
PySocks==1.7.1
pytest==5.3.5
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rope==0.16.0
selenium==3.141.0
tables==3.6.1
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
unicodecsv==0.14.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
win-inet-pton==1.1.0
win-unicode-console==0.5